refactor: create framer interface
