+++ internal/core/models/models.go
@@ -2,10 +2,9 @@ package models
 
 import (
 	"github.com/andrelcunha/ottermq/internal/core/amqp"
-	"github.com/andrelcunha/ottermq/internal/core/amqp/shared"
 )
 
 type ConnectionInfo struct {
-	Client   *shared.AmqpClient
+	Client   *amqp.AmqpClient
 	Channels map[uint16]*amqp.ChannelState `json:"-"`
 }
